sanity test on MNIST dataset

+++ stochnet_v2/dynamic_classes/util.py
@@ -13,6 +13,11 @@ def preprocess(layer_input):
     return layer_preprocess(layer_input, 'none', 0.1, 'layer', 1e-4)  # 'none'
 
 
+# def postprocess(layer_output):
+#     return layer_postprocess(0., layer_output, 'n', 0.1, 'l2', 1e-4)  # 'n'
+
+
+# TODO: for MNIST only:
 def postprocess(layer_output):
     return layer_postprocess(0., layer_output, 'n', 0.1, 'l2', 1e-4)  # 'n'
 
@@ -128,7 +133,7 @@ def _simple_dense(
         activity_regularizer=None,
         **kwargs,
 ):
-    n_units = x.shape.as_list()[-1] * expansion_coeff
+    n_units = int(x.shape.as_list()[-1] * expansion_coeff)
 
     with tf.variable_scope("Dense"):
         x = preprocess(x)
@@ -162,7 +167,7 @@ def _activated_dense(
         activity_regularizer=None,
         **kwargs,
 ):
-    n_units = x.shape.as_list()[-1] * expansion_coeff
+    n_units = int(x.shape.as_list()[-1] * expansion_coeff)
 
     with tf.variable_scope("ActivatedDense"):
         x = preprocess(x)
@@ -197,7 +202,7 @@ def _gated_linear_unit(
         activity_regularizer=None,
         **kwargs,
 ):
-    n_units = x.shape.as_list()[-1] * expansion_coeff
+    n_units = int(x.shape.as_list()[-1] * expansion_coeff)
 
     with tf.variable_scope("GatedLinearUnit"):
         if expansion_coeff > 1:
